MySQL Database Project Ready Added all the components necessary to work with each tables (Add, Remove, Get)

+++ src/scripts/create_schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Country(
    CId VARCHAR(3),
    name VARCHAR(50) NOT NULL,
-   gmtOffset INT NOT NULL,
+   gmtOffset DECIMAL(2,1) NOT NULL,
    PRIMARY KEY(CId)
 );
